# Conflicts:
#	src/main/java/GUI/DrawGuitarBeams.java

+++ src/main/java/GUI/DrawGuitarBeams.java
@@ -27,10 +27,10 @@ public class DrawGuitarBeams {
 	 */
 	private ArrayList<Note> notes, noteWithoutChords;
 	private ArrayList<NoteLocation> noteLocations, notesWithoutChordsLocations;
-	private int noteSpacing;
+	private double noteSpacing;
 	private Pane pane;
 	
-	public DrawGuitarBeams(Pane pane, ArrayList<Note> notes, ArrayList<NoteLocation> noteLocations, int noteSpacing) {
+	public DrawGuitarBeams(Pane pane, ArrayList<Note> notes, ArrayList<NoteLocation> noteLocations, double noteSpacing) {
 		this.pane = pane;
 		this.notes = notes;
 		this.noteLocations = noteLocations;
@@ -286,18 +286,18 @@ public class DrawGuitarBeams {
 	}
 	
 
-	public void draw8thBeam(double startX, double startY, int length) {
+	public void draw8thBeam(double startX, double startY, double length) {
 		Rectangle r = new Rectangle(startX, startY, length, 4);
 		pane.getChildren().add(r);
 	}
 	
-	public void draw16thBeam(double startX, double startY, int length) {
+	public void draw16thBeam(double startX, double startY, double length) {
 		draw8thBeam(startX, startY, length);
 		Rectangle r = new Rectangle(startX, startY-7, length, 4);
 		pane.getChildren().add(r);
 	}
 	
-	public void draw32ndBeam(double startX, double startY, int length) {
+	public void draw32ndBeam(double startX, double startY, double length) {
 		draw16thBeam(startX, startY, length);
 		Rectangle r = new Rectangle(startX, startY-14, length, 4);
 		pane.getChildren().add(r);
